requirements.txt
langchain
langchain-community
langchain-core
langgraph
langgraph-cli
langgraph-api
langchain-groq
langchain-huggingface
groq
pypdf
pypdfium2
langchain-text-splitters
chromadb
sentence-transformers
huggingface-hub
numpy
pydantic
requests
typing-extensions
python-dotenv